CSS + Program start
Fixed some basic css for the homepage, login page and gamepage. The game page will be changed when able to test with images and all database data. Program cs: fixed where the program starts. Also fixed that the user will be redirected to the login page automatically when opening webpage